+++ backend/DTOs/ActivitySessionDetailDto.cs
@@ -16,7 +16,7 @@ namespace backend.DTOs
 
         public TimeOnly? EndTime { get; set; }
 
-        public int? DaysCount { get; set; }
+        public decimal? DaysCount { get; set; }
 
         public int? MaxParticipants { get; set; }
